Move all uuid methods into game-profiles.ts
