+++ src/index.ts
@@ -22,9 +22,9 @@ import cors from "@elysiajs/cors";
 import { verify as verifyMailOptions } from "./libs/mailer";
 import { getLatestCommit, retrieveData } from "./libs/git-commit-data";
 import { startEntitlementExpiry, startMetrics, startReferralReset, startRoleCacheJob } from "./libs/cron-jobs";
-import { formatUUID } from "./routes/players/[uuid]/root";
 import { config } from "./libs/config";
 import { join } from "path";
+import { formatUUID } from "./libs/game-profiles";
 
 if(config.mongodb.trim().length == 0) {
     Logger.error('Database connection string is empty!');
