+++ src/libs/game-profiles.ts
@@ -53,3 +53,15 @@ export async function getProfileByUsername(username: string): Promise<Profile> {
         }
     }
 }
+
+export function formatUUID(uuid: string): string {
+    const cleanedUUID = uuid.replace(/-/g, '');
+    
+    if(cleanedUUID.length != 32) throw new Error("Invalid UUID length: Expected 32 characters without dashes.");
+    
+    return `${cleanedUUID.slice(0, 8)}-${cleanedUUID.slice(8, 12)}-${cleanedUUID.slice(12, 16)}-${cleanedUUID.slice(16, 20)}-${cleanedUUID.slice(20)}`;
+}
+
+export function stripUUID(uuid: string): string {
+    return uuid.replace(/-/g, '');
+}
