+++ src/routes/players/[uuid]/ban.ts
@@ -5,13 +5,13 @@ import { ModLogType, sendBanAppealMessage, sendModLogMessage } from "../../../li
 import getAuthProvider from "../../../middleware/get-auth-provider";
 import { sendBanEmail, sendUnbanEmail } from "../../../libs/mailer";
 import { Permission } from "../../../types/Permission";
-import { getProfileByUUID } from "../../../libs/game-profiles";
+import { getProfileByUUID, stripUUID } from "../../../libs/game-profiles";
 
 export default new Elysia({
     prefix: `/ban`
 }).use(fetchI18n).use(getAuthProvider).get(`/`, async ({ error, params, headers, i18n, provider }) => { // Get ban info
     if(!provider) return error(401, { error: i18n('error.malformedAuthHeader') });
-    const uuid = params.uuid.replaceAll(`-`, ``);
+    const uuid = stripUUID(params.uuid);
     const { authorization } = headers;
     const session = await provider.getSession(authorization, uuid);
     if(!session.hasPermission(Permission.ManageBans)) return error(403, { error: i18n(`error.notAllowed`) });
@@ -35,7 +35,7 @@ export default new Elysia({
     headers: t.Object({ authorization: t.String({ error: `error.notAllowed`, description: `Your LabyConnect JWT` }) }, { error: `error.notAllowed` })
 }).post(`/`, async ({ error, params, headers, body, i18n, provider }) => { // Ban player
     if(!provider) return error(401, { error: i18n('error.malformedAuthHeader') });
-    const uuid = params.uuid.replaceAll(`-`, ``);
+    const uuid = stripUUID(params.uuid);
     const { authorization } = headers;
     const session = await provider.getSession(authorization, uuid);
     if(!session.hasPermission(Permission.ManageBans)) return error(403, { error: i18n(`error.notAllowed`) });
@@ -78,7 +78,7 @@ export default new Elysia({
     headers: t.Object({ authorization: t.String({ error: `error.notAllowed`, description: `Your LabyConnect JWT` }) }, { error: `error.notAllowed` })
 }).put(`/`, async ({ error, params, headers, body, i18n, provider }) => { // Update ban info - I need to use put here bc labymod's Request system doesn't support PATCH
     if(!provider) return error(401, { error: i18n('error.malformedAuthHeader') });
-    const uuid = params.uuid.replaceAll(`-`, ``);
+    const uuid = stripUUID(params.uuid);
     const { authorization } = headers;
     const session = await provider.getSession(authorization, uuid);
     if(!session.hasPermission(Permission.ManageBans)) return error(403, { error: i18n(`error.notAllowed`) });
@@ -119,7 +119,7 @@ export default new Elysia({
     headers: t.Object({ authorization: t.String({ error: `error.notAllowed`, description: `Your LabyConnect JWT` }) }, { error: `error.notAllowed` })
 }).post(`/appeal`, async ({ error, params, headers, body: { reason }, i18n, provider }) => { // Ban player
     if(!provider) return error(401, { error: i18n('error.malformedAuthHeader') });
-    const uuid = params.uuid.replaceAll(`-`, ``);
+    const uuid = stripUUID(params.uuid);
     const { authorization } = headers;
     const session = await provider.getSession(authorization, uuid);
     if(!session.equal) return error(403, { error: i18n(`error.notAllowed`) });
@@ -154,7 +154,7 @@ export default new Elysia({
     headers: t.Object({ authorization: t.String({ error: `error.notAllowed`, description: `Your LabyConnect JWT` }) }, { error: `error.notAllowed` })
 }).delete(`/`, async ({ error, params, headers, i18n, provider }) => { // Unban player
     if(!provider) return error(401, { error: i18n('error.malformedAuthHeader') });
-    const uuid = params.uuid.replaceAll(`-`, ``);
+    const uuid = stripUUID(params.uuid);
     const { authorization } = headers;
     const session = await provider.getSession(authorization, uuid);
     if(!session.hasPermission(Permission.ManageBans)) return error(403, { error: i18n(`error.notAllowed`) });
